+++ Mods ChoGGi/ChoGGi Decorations/metadata.lua
@@ -1,0 +1,23 @@
+return PlaceObj("ModDef", {
+	"title", "ChoGGi Decorations",
+	"id", "ChoGGi_ChoGGiDecorations",
+	"steam_id", "2474720549",
+	"pops_any_uuid", "9cc6412d-ea55-488a-bd91-174f661d2470",
+	"lua_revision", 1001514, -- Tito
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.png",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+--~ 	"has_options", true,
+	"TagBuildings", true,
+	"description", [[
+Does nothing, looks pretty.
+
+
+Requested by CanadianACOG.
+]],
+})