+++ Mods ChoGGi/Rocket Cheats/Code/Script.lua
@@ -28,19 +28,15 @@ function RocketBase:FlyToMars(cargo, cost, flight_time, ...)
 end
 
 -- some stuff checks one some other...
---~ local SetConsts = ChoGGi.ComFuncs.SetConstsG
-local function SetConsts(id, value)
-	Consts[id] = value
-	g_Consts[id] = value
-end
+local SetConstsG = ChoGGi.ComFuncs.SetConstsG
 
 local function UpdateRockets()
-	SetConsts("MaxColonistsPerRocket", mod_MaxColonistsPerRocket)
-	SetConsts("TravelTimeEarthMars", mod_TravelTimeEarthMars)
-	SetConsts("TravelTimeMarsEarth", mod_TravelTimeMarsEarth)
-	SetConsts("RocketPrice", mod_RocketPrice)
-	SetConsts("CargoCapacity", mod_CargoCapacity)
-	SetConsts("FoodPerRocketPassenger", mod_FoodPerRocketPassenger)
+	SetConstsG("MaxColonistsPerRocket", mod_MaxColonistsPerRocket)
+	SetConstsG("TravelTimeEarthMars", mod_TravelTimeEarthMars)
+	SetConstsG("TravelTimeMarsEarth", mod_TravelTimeMarsEarth)
+	SetConstsG("RocketPrice", mod_RocketPrice)
+	SetConstsG("CargoCapacity", mod_CargoCapacity)
+	SetConstsG("FoodPerRocketPassenger", mod_FoodPerRocketPassenger)
 
 	local objs =  UICity.labels.SupplyRocket or ""
 	for i = 1, #objs do